Move components in theatres directory to top level
