+++ packages/ia-components/index.js
@@ -9,8 +9,8 @@ export { default as IAUXExampleComponent } from './live/example-react-component/
 export { default as HorizontalRadioGroup } from './sandbox/selectors/horizontal-radio-group/horizontal-radio-group';
 export { default as FlexboxPagination } from './sandbox/pagination/pagination-with-flexbox';
 
-export { default as TheatreTrackList } from './sandbox/theatres/components/track-list/track-list';
-export { default as TheatreAudioPlayer } from './sandbox/theatres/components/audio-player/audio-player-main';
+export { default as TheatreTrackList } from './sandbox/theatres/track-list/track-list';
+export { default as TheatreAudioPlayer } from './sandbox/theatres/audio-player/audio-player-main';
 export { default as AudioPlayerWithYoutubeSpotify } from './sandbox/theatres/with-youtube-spotify/audio-with-youtube-spotify-main';
 export { default as AnchorDetails } from './sandbox/AnchorDetails';
 // sandbox tiles
@@ -24,4 +24,4 @@ export { default as DetailsActionButtons } from './sandbox/details/DetailsAction
 export { default as DetailsDownloadOptions } from './sandbox/details/DetailsDownloadOptions';
 export { default as DetailsFlags } from './sandbox/details/DetailsFlags';
 export { AnchorModalGo, ButtonModalGo } from './sandbox/details/ModalGo';
-export { default as Tabby} from './sandbox/details/Tabby';
+export { default as Tabby } from './sandbox/details/Tabby';
