+++ packages/ia-components/index.less
@@ -3,7 +3,7 @@
 @import './live/example-react-component/styles.less';
 
 @import './sandbox/theatres/with-youtube-spotify/audio-with-youtube-spotify.less';
-@import './sandbox/theatres/components/track-list/track-list.less';
-@import './sandbox/theatres/components/audio-player/audio-player.less';
+@import './sandbox/theatres/track-list/track-list.less';
+@import './sandbox/theatres/audio-player/audio-player.less';
 @import './sandbox/selectors/horizontal-radio-group/horizontal-radio-group.less';
 @import './sandbox/pagination/pagination.less';
