+++ packages/ia-components/sandbox/theatres/svgs/ia-logo-white-icon.js
