+++ packages/ia-components/sandbox/theatres/svgs/icon-booklet.js
