+++ packages/ia-components/sandbox/theatres/svgs/spotify-logo-icon.js
