+++ packages/ia-components/sandbox/theatres/svgs/youtube-logo-icon.js
