+++ packages/ia-components/sandbox/theatres/track-list/track-list.less
@@ -1,11 +1,11 @@
-@import '../../../../variables.less';
+@import '../../../variables.less';
 
 .audio-track-list {
   position: relative;
   // split overflow declaration needed for Safari
   overflow-x: hidden;
   overflow-y: auto;
-  
+
   .page-buttons {
     .pagination-button {
       border-color: #fff;
@@ -34,7 +34,7 @@
     &.selected {
       border-color:#00A1DD;
       border-radius: .5rem;
-  
+
       &:focus {
         outline: none;
       }
